split responsibilities
